@@ -20,39 +20,40 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@codeceptjs/configure": "^0.7.0",
-    "@codeceptjs/ui": "^0.4.6",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.1.2",
-    "@testing-library/react-hooks": "^7.0.2",
-    "@types/jest": "^27.0.2",
-    "@types/react": "^17.0.27",
-    "@types/react-dom": "^17.0.9",
+    "@codeceptjs/configure": "^0.10.0",
+    "@codeceptjs/ui": "^0.4.7",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@types/jest": "^29.2.4",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.10",
     "@types/recoil": "^0.0.9",
-    "@types/styled-components": "^5.1.15",
-    "@typescript-eslint/eslint-plugin": "^5.0.0",
-    "@typescript-eslint/parser": "^5.0.0",
-    "codeceptjs": "^3.1.3",
-    "concurrently": "^6.3.0",
-    "esbuild": "^0.13.4",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb": "^18.2.1",
+    "@types/styled-components": "^5.1.26",
+    "@typescript-eslint/eslint-plugin": "^5.47.0",
+    "@typescript-eslint/parser": "^5.47.0",
+    "codeceptjs": "^3.3.7",
+    "concurrently": "^7.6.0",
+    "esbuild": "^0.16.10",
+    "eslint": "^8.30.0",
+    "eslint-config-airbnb": "^19.0.4",
     "eslint-plugin-codeceptjs": "^1.3.0",
-    "eslint-plugin-import": "^2.24.2",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.26.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "jest": "^27.2.5",
-    "playwright": "^1.15.2",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "^6.6.1",
+    "eslint-plugin-react": "^7.31.11",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "jest": "^29.3.1",
+    "jest-environment-jsdom": "^29.3.1",
+    "playwright": "^1.29.1",
     "servor": "^4.0.2",
-    "ts-jest": "^27.0.7",
-    "typescript": "^4.4.4"
+    "ts-jest": "^29.0.3",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.4"
   },
   "dependencies": {
-    "axios": "^0.22.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "recoil": "^0.4.1",
-    "styled-components": "^5.3.1"
+    "axios": "^1.2.1",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "recoil": "^0.7.6",
+    "styled-components": "^5.3.6"
   }
 }
