@@ -1,12 +1,17 @@
+/* eslint-disable import/prefer-default-export */
+
 // eslint-disable-next-line import/no-extraneous-dependencies
-const { setHeadlessWhen } = require('@codeceptjs/configure');
+import { setHeadlessWhen, setCommonPlugins } from '@codeceptjs/configure';
 
 // turn on headless mode when running with HEADLESS=true environment variable
 // export HEADLESS=true && npx codeceptjs run
 setHeadlessWhen(process.env.HEADLESS);
 
-exports.config = {
-  tests: './tests/**/*_test.js',
+// enable all common plugins https://github.com/codeceptjs/configure#setcommonplugins
+setCommonPlugins();
+
+export const config = {
+  tests: './tests/**/*_test.ts',
   output: './output',
   helpers: {
     Playwright: {
@@ -15,19 +20,9 @@ exports.config = {
       browser: 'chromium',
     },
   },
-  bootstrap: null,
-  mocha: {},
-  name: 'react',
-  plugins: {
-    pauseOnFail: {},
-    retryFailedStep: {
-      enabled: true,
-    },
-    tryTo: {
-      enabled: true,
-    },
-    screenshotOnFail: {
-      enabled: true,
-    },
+  include: {
+    I: './tests/steps_file',
   },
+  name: 'react',
+  fullPromiseBased: false,
 };
